+++ core/embed/io/touch/ft6x36/ft6x36.c
@@ -239,7 +239,7 @@ static secbool ft6x36_configure(i2c_bus_t* i2c_bus) {
 
   _Static_assert(sizeof(config) % 2 == 0);
 
-  for (int i = 0; i < sizeof(config); i += 2) {
+  for (size_t i = 0; i < sizeof(config); i += 2) {
     uint8_t reg = config[i];
     uint8_t value = config[i + 1];
