+++ core/embed/sys/startup/inc/sys/bootargs.h
@@ -43,7 +43,7 @@ typedef enum {
 } boot_command_t;
 
 // Maximum size boot_args array
-#define BOOT_ARGS_MAX_SIZE (256 - 8)
+#define BOOT_ARGS_MAX_SIZE ((size_t)(256 - 8))
 
 typedef union {
   uint8_t raw[BOOT_ARGS_MAX_SIZE];
